feat: improved the layout of the home page
