+++ package.json
@@ -22,6 +22,7 @@
     "@vercel/speed-insights": "^1.0.12",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
+    "framer-motion": "^11.3.29",
     "lucide-react": "^0.354.0",
     "next": "14.1.2",
     "next-themes": "^0.3.0",
